refactor: UI로 인한 의존성 사이클을 해결한다
