+++ src/main/java/kitchenpos/menu/dto/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.menu.dto;
 
 public class MenuProductRequest {
