+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.order.dto;
 
 import java.util.List;
