+++ src/main/java/kitchenpos/ordertable/dto/OrderStatusChangeRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ordertable.dto;
 
 import kitchenpos.order.OrderStatus;
