+++ src/main/java/kitchenpos/ordertable/dto/OrderTableChangeEmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ordertable.dto;
 
 public class OrderTableChangeEmptyRequest {
