+++ src/main/java/kitchenpos/ordertable/dto/OrderTableResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ordertable.dto;
 
 import kitchenpos.ordertable.OrderTable;
