+++ src/main/java/kitchenpos/tablegroup/dto/OrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.tablegroup.dto;
 
 public class OrderTableRequest {
