+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.ui;
 
 import kitchenpos.menugroup.MenuGroup;
 import kitchenpos.menugroup.MenuGroupService;
-import kitchenpos.ui.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
