+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.ui;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderService;
 import kitchenpos.order.OrderStatus;
-import kitchenpos.ui.dto.OrderRequest;
-import kitchenpos.ui.dto.OrderStatusChangeRequest;
+import kitchenpos.order.dto.OrderRequest;
+import kitchenpos.ordertable.dto.OrderStatusChangeRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
