+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.ui;
 
 import kitchenpos.product.Product;
 import kitchenpos.product.ProductService;
-import kitchenpos.ui.dto.ProductRequest;
+import kitchenpos.product.dto.ProductRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
