+++ src/test/java/kitchenpos/ordertable/TableServiceTest.java
@@ -4,11 +4,11 @@ import kitchenpos.fixture.OrderTableFixture;
 import kitchenpos.menu.Menu;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderService;
+import kitchenpos.order.dto.OrderLineItemRequest;
+import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.tablegroup.TableGroupService;
-import kitchenpos.ui.dto.OrderLineItemRequest;
-import kitchenpos.ui.dto.OrderRequest;
-import kitchenpos.ui.dto.OrderTableRequest;
-import kitchenpos.ui.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.OrderTableRequest;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
