+++ src/test/java/kitchenpos/tablegroup/TableGroupServiceTest.java
@@ -3,12 +3,12 @@ package kitchenpos.tablegroup;
 import kitchenpos.menu.Menu;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderService;
+import kitchenpos.order.dto.OrderLineItemRequest;
+import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.ordertable.OrderTable;
 import kitchenpos.ordertable.TableService;
-import kitchenpos.ui.dto.OrderLineItemRequest;
-import kitchenpos.ui.dto.OrderRequest;
-import kitchenpos.ui.dto.OrderTableRequest;
-import kitchenpos.ui.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.OrderTableRequest;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
